7119における発熱等相談件数: Rename components/cards/ConsultationAboutFeverNumberCard.vue to components/index/CardsMonitoring/ConsultationAboutFeverNumber/Card.vue
