+++ components/index/CardsMonitoring/ConsultationAboutFeverNumber/Card.vue
