+++ libraries/repositories/ConsultationAboutFeverRepository.ts
@@ -19,7 +19,7 @@ export class ConsultationAboutFeverRepository
   /**
    * 使用箇所
    *
-   * モニタリング項目(2) #7119における発熱等相談件数 (components/cards/ConsultationAboutFeverNumberCard.vue)
+   * モニタリング項目(2) #7119における発熱等相談件数 (components/index/CardsMonitoring/ConsultationAboutFeverNumber/Card.vue)
    */
   constructor() {
     super(Convert.toConsultationAboutFever(JSON.stringify(rawData)))
